requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Artem\image-filter\image_classification.py: 3,4
# C:\Users\Artem\image-filter\main.py: 1
Pillow == 8.1.0

# C:\Users\Artem\image-filter\image_classification.py: 2
# C:\Users\Artem\image-filter\main.py: 2
numpy == 1.20.1

# C:\Users\Artem\image-filter\text_extraction.py: 1
pytesseract == 0.3.7

# C:\Users\Artem\image-filter\image_classification.py: 5
# C:\Users\Artem\image-filter\main.py: 3
requests == 2.25.1

# C:\Users\Artem\image-filter\image_classification.py: 1
tensorflow == 2.4.1
